backend/services/system_architect_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
PyYAML==6.0.1
networkx==3.2.1
aiofiles==23.2.1
